requirements.txt
boto3
pandas
sqlalchemy
psycopg2-binary
google-cloud-storage
faker
python-dotenv